chore: rename winner ribbon
